changed file name form dommanipulation to renderers, since they are rendering to the page
