+++ index.html
@@ -58,6 +58,6 @@
                 <img class="category">
             </div>
         </template>
-
+        <script src="script.js"></script>
     </body>
 </html>
