+++ tests.js
@@ -1,23 +1,26 @@
-//Simon - test 1 
+//Simon - test 1
+// import {} from './handlers'
+// import {} from './script'
+// import {} from './renderers'
 
 test("Submitting a new task adds it to the list", () => {
   // Get the object for the values inputted and find the tag to check against the created To Do
-  
+
   const inputObject = submitNewToDo();
   const tagToTest = inputObject.tag;
 
   // Reassign the data from the created To Do to an object in the same format as the values inputted object
-  
-  const toDoToTest = toDoContainer.querySelector(`#${tagToTest}`);
-  const iconForCategoryToTest = toDoToTest.querySelector('img');
+
+  const toDoTest = toDoContainer.querySelector(`#${tagToTest}`);
+  const iconForCategoryToTest = toDoTest.querySelector("img");
   const categoryToTest = iconForCategoryToTest.id;
-  const ouputObject = {
-    text: toDoToTest.querySelector("#textContent"),
+  const outputObject = {
+    text: toDoTest.querySelector("#textContent"),
     category: categoryToTest,
-    date: toDoToTest.querySelector("#date"),
-    time: toDoToTest.querySelector("#time"),
-    tag: toDoTest.id 
-  }
+    date: toDoTest.querySelector("#date"),
+    time: toDoTest.querySelector("#time"),
+    tag: toDoTest.id,
+  };
 
   // Carry out checks\
 
@@ -28,5 +31,4 @@ test("Submitting a new task adds it to the list", () => {
   equal(result.date, expected.date);
   equal(result.time, expected.time);
   equal(result.tag, expected.tag);
-
 });
